More bug fixings. Everything works now and it is time for bed.
git-svn-id: http://code.icecube.wisc.edu/svn/projects/photospline/trunk@89093 16731396-06f5-0310-8873-f7f720988828

+++ private/lib/bspline.c
@@ -418,7 +418,7 @@ ndsplineeval(struct splinetable *table, const double *x, const int *centers,
     int derivatives)
 {
 	int i, j, n, tablepos;
-	double result;
+	float result;
 	int maxdegree = maxorder(table->order, table->ndim) + 1; 
 	float localbasis[table->ndim][maxdegree];
 	float basis_tree[table->ndim+1];
